split achievement badge component
